+++ apps/alu-learn/helpers/getServerSession.ts
@@ -4,7 +4,7 @@ import type {
   NextApiResponse,
   PreviewData,
 } from "next";
-import { unstable_getServerSession } from "next-auth";
+import { getServerSession as nextAuth_getServerSession } from "next-auth";
 import { authOptions } from "pages/api/auth/[...nextauth]";
 import type { ParsedUrlQuery } from "querystring";
 
@@ -12,6 +12,6 @@ const getServerSession = (
   context:
     | GetServerSidePropsContext<ParsedUrlQuery, PreviewData>
     | { req: NextApiRequest; res: NextApiResponse<any> }
-) => unstable_getServerSession(context.req, context.res, authOptions);
+) => nextAuth_getServerSession(context.req, context.res, authOptions);
 
 export default getServerSession;